+++ glassfish-runner/enterprise-beans-tck/enterprise-beans-tck-run/src/test/java/arquillian/GlassFishTckExtension.java
@@ -1,0 +1,12 @@
+package arquillian;
+
+import org.jboss.arquillian.core.spi.LoadableExtension;
+import org.jboss.arquillian.test.spi.enricher.resource.ResourceProvider;
+
+public class GlassFishTckExtension implements LoadableExtension {
+    @Override
+    public void register(ExtensionBuilder builder) {
+        builder.service(ResourceProvider.class, GlassFishXmlProcessor.class);
+        builder.observer(GlassFishXmlProcessor.class);
+    }
+}
