+++ glassfish-runner/enterprise-beans-tck/enterprise-beans-tck-run/src/test/java/arquillian/ProgressListener.java
@@ -1,8 +1,6 @@
 package arquillian;
 
-import java.util.ArrayList;
 import java.util.LinkedHashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
