+++ glassfish-runner/enterprise-beans-tck/enterprise-beans-tck-run/src/test/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,1 @@
-wildfly.arquillian.JBossTckExtension
+arquillian.GlassFishTckExtension
