--- glassfish-runner/enterprise-beans-tck/enterprise-beans-tck-run/src/test/resources/ejb30/assembly/initorder/warejb/ejb3_assembly_initorder_warejb_ejb.jar.jboss-ejb3.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jboss:jboss xmlns:jboss="http://www.jboss.com/xml/ns/javaee"
-             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xsi:schemaLocation="http://www.jboss.com/xml/ns/javaee http://www.jboss.org/j2ee/schema/jboss_5_0.xsd http://java.sun.com/xml/ns/javaee http://www.jboss.org/j2ee/schema/jboss-ejb3-spec-2_0.xsd urn:security jboss-ejb-security_1_0.xsd urn:security-role jboss-ejb-security-role_1_0.xsd"
-             version="3.1" impl-version="2.0">
-    <jboss:enterprise-beans>
-        <jboss:ejb>
-            <ejb-name xmlns="http://java.sun.com/xml/ns/javaee">InitOrderBean</ejb-name>
-            <activation-config xmlns="http://java.sun.com/xml/ns/javaee">
-                <activation-config-property>
-                    <activation-config-property-name>destination</activation-config-property-name>
-                    <activation-config-property-value>java:/ejb3_assembly_initorder_warejb_InitOrderBean</activation-config-property-value>
-                </activation-config-property>
-            </activation-config>
-        </jboss:ejb>
-    </jboss:enterprise-beans>
-    <assembly-descriptor xmlns="http://java.sun.com/xml/ns/javaee">
-        <iiop xmlns:xalan="http://xml.apache.org/xalan" xmlns:sr="urn:security-role" xmlns:s="urn:security:1.1" xmlns:ext="xalan://org.jboss.ts.lib.porting.EjbJarHandler" xmlns="urn:iiop">
-            <ejb-name xmlns="http://java.sun.com/xml/ns/javaee">InitOrderBean</ejb-name>
-            <binding-name>ejb3_assembly_initorder_warejb_InitOrderBean</binding-name>
-        </iiop>
-    </assembly-descriptor>
-</jboss:jboss>
