--- glassfish-runner/enterprise-beans-tck/enterprise-beans-tck-run/src/test/resources/ejb30/bb/session/stateless/basic/ejb3_bb_stateless_basic_ejb.jar.sun-ejb-jar.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE sun-ejb-jar PUBLIC "-//Sun Microsystems, Inc.//DTD Application Server 9.0 EJB 3.0//EN" "http://www.sun.com/software/appserver/dtds/sun-ejb-jar_3_0-0.dtd">
-<!--
-
-    Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0, which is available at
-    http://www.eclipse.org/legal/epl-2.0.
-
-    This Source Code may also be made available under the following Secondary
-    Licenses when the conditions for such availability set forth in the
-    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
-    version 2 with the GNU Classpath Exception, which is available at
-    https://www.gnu.org/software/classpath/license.html.
-
-    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-
--->
-
-<sun-ejb-jar>
-    <enterprise-beans>
-    <ejb>
-      <ejb-name>RemoteCalculatorBean</ejb-name>
-      <jndi-name>com_sun_ts_tests_ejb30_bb_session_stateless_basic_RemoteCalculatorBean</jndi-name>
-    </ejb>
-    <ejb>
-      <ejb-name>RemoteCalculatorBean2</ejb-name>
-      <jndi-name>com_sun_ts_tests_ejb30_bb_session_stateless_basic_RemoteCalculatorBean2</jndi-name>
-    </ejb>
-    <ejb>
-      <ejb-name>RemoteCalculatorBean3</ejb-name>
-      <jndi-name>com_sun_ts_tests_ejb30_bb_session_stateless_basic_RemoteCalculatorBean3</jndi-name>
-    </ejb>
-    <ejb>
-      <ejb-name>RemoteCalculatorBean4</ejb-name>
-      <jndi-name>com_sun_ts_tests_ejb30_bb_session_stateless_basic_RemoteCalculatorBean4</jndi-name>
-    </ejb>
-    <ejb>
-      <ejb-name>RemoteCalculatorBean5</ejb-name>
-      <jndi-name>com_sun_ts_tests_ejb30_bb_session_stateless_basic_RemoteCalculatorBean5</jndi-name>
-    </ejb>
-    </enterprise-beans>
-</sun-ejb-jar>
