--- glassfish-runner/enterprise-beans-tck/enterprise-beans-tck-run/src/test/resources/ejb30/lite/view/singleton/annotated/ejbliteservlet_vehicle_web.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0, which is available at
-    http://www.eclipse.org/legal/epl-2.0.
-
-    This Source Code may also be made available under the following Secondary
-    Licenses when the conditions for such availability set forth in the
-    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
-    version 2 with the GNU Classpath Exception, which is available at
-    https://www.gnu.org/software/classpath/license.html.
-
-    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-
--->
-
-<web-app version="5.0" xmlns="https://jakarta.ee/xml/ns/jakartaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee https://jakarta.ee/xml/ns/jakartaee/web-app_5_0.xsd">
-    <servlet>
-        <servlet-name>EJBLiteServletVehicle</servlet-name>
-        <servlet-class>ejb30.lite.view.singleton.annotated.EJBLiteServletVehicle</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>EJBLiteServletVehicle</servlet-name>
-        <url-pattern>/ejbliteservlet_vehicle.jsp</url-pattern>
-    </servlet-mapping>
-</web-app>
